+++ Packages/com.mimylab.fukuroudon/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 利用手順は[こちら](https://github.com/mimyquality/FukuroUdon/wiki)からご確認ください。
 
+## [3.13.0-beta.3] - 2025/11/3
+
+- **Added**
+  - Dynamics Parameter Transfer に [Setup Contacts in Children] ボタンを追加しました。
+    - 子孫にある全ての VRC Contact Receiver に Contact Receiver Infomation を追加するボタンです。
+
 ## [3.13.0-beta.2] - 2025/11/2
 
 - **Added**
@@ -758,6 +764,7 @@
     - 0.2.7バージョンで、later-joiner視点でjoin直後が非アクティブだと同期しなかったのを修正しました。
     - 他人がピックアップした際にオブジェクトが一瞬跳ねる現象を低減しました。
 
+[3.13.0-beta.3]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.13.0-beta.3
 [3.13.0-beta.2]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.13.0-beta.2
 [3.13.0-beta.1]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.13.0-beta.1
 [3.12.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/3.12.0